requirements.txt
feedparser==6.0.11
python-dotenv==1.0.1
unidecode==1.3.8
newspaper3k==0.2.8
lxml_html_clean==0.1.0
langchain==0.1.14
text_generation==0.7.0